Stop rendering when camera is still

+++ shaders/blurFrag.glsl
@@ -7,7 +7,7 @@ uniform highp vec2 shift;
 uniform highp float sizeShift;
 
 void main() {
-    lowp vec3 texval = texture2D(tex, pos).xyz;
+    highp vec3 texval = texture2D(tex, pos).xyz;
 
     highp vec2 shiftedUV = (pos + shift - vec2(0.5, 0.5)) * (1.0 - sizeShift) + vec2(0.5, 0.5);
 
@@ -17,12 +17,8 @@ void main() {
     }
     else
     {
-        lowp vec3 cumulval = texture2D(cumul, shiftedUV).xyz;
-
-        // texval.x += sin(pos.x * 180.0) * sin(pos.x * 180.0) * sin(pos.x * 180.0);
+        highp vec3 cumulval = texture2D(cumul, shiftedUV).xyz;
         
         gl_FragColor = vec4(texval * 0.05 + cumulval * 0.95, 1.0);
-
-        //gl_FragColor = vec4(texval * 0.005 + cumulval * 0.995, 1.0);
     }
 }
